Address possible loss of data warnings
From Visual Studio, for example:

_deps\cleaver_lib-src\src\lib\cleaver\CleaverMesher.cpp(160): warning C4018: '<': signed/unsigned mismatch
flag
Build Log Line	24
Warning
_deps\cleaver_lib-src\src\lib\cleaver\CleaverMesher.cpp(316): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
flag
Build Log Line	25
Warning
_deps\cleaver_lib-src\src\lib\cleaver\CleaverMesher.cpp(433): warning
C4244: 'return': conversion from 'double' to 'float', possible loss of
data

+++ src/lib/cleaver/Debug.cpp
@@ -77,9 +77,9 @@ namespace cleaver
 
   std::string idForEdge(HalfEdge *edge) {
     Vertex *v1 = edge->vertex;
-    Vertex *v2 = edge->mate->vertex;    
+    Vertex *v2 = edge->mate->vertex;
     if (v1->tm_v_index > v2->tm_v_index) {
-      std::swap(v1, v2);      
+      std::swap(v1, v2);
     }
     return catIds({ v1->tm_v_index, v2->tm_v_index});
   }
@@ -102,23 +102,23 @@ namespace cleaver
     Vertex *v1 = edge->vertex;
     Vertex *v2 = edge->mate->vertex;
     double alpha1 = edge->alpha;
-    double alpha2 = edge->mate->alpha;    
+    double alpha2 = edge->mate->alpha;
     if (v1->tm_v_index > v2->tm_v_index) {
       std::swap(v1, v2);
-      std::swap(alpha1, alpha2);      
+      std::swap(alpha1, alpha2);
     }
 
     Json::Value root(Json::objectValue);
     root["name"] = "CREATE_EDGE";
     root["id"] = catIds({ v1->tm_v_index, v2->tm_v_index}).c_str();
-    root["v1"] = v1->tm_v_index;
-    root["v2"] = v2->tm_v_index;
+    root["v1"] = static_cast<int>(v1->tm_v_index);
+    root["v2"] = static_cast<int>(v2->tm_v_index);
     root["alpha1"] = alpha1;
     root["alpha2"] = alpha2;
 
     if (edge->cut && edge->cut->order() == Order::CUT) {
       // TODO(jonbronson): set a UID and add operation to create separately
-      root["cut"] = createVertexOperation(edge->cut);  
+      root["cut"] = createVertexOperation(edge->cut);
       root["cut"]["violating"] = edge->cut->violating;
     }
 
@@ -133,8 +133,8 @@ namespace cleaver
     Vertex *v3 = face->halfEdges[2]->vertex;
 
     std::vector<Vertex*> vertexList = {v1, v2, v3};
-    std::sort(vertexList.begin(), vertexList.end(), [](Vertex* a, Vertex* b) -> bool 
-      { 
+    std::sort(vertexList.begin(), vertexList.end(), [](Vertex* a, Vertex* b) -> bool
+      {
         return (a->tm_v_index < b->tm_v_index);
       });
     v1 = vertexList[0];
@@ -144,9 +144,9 @@ namespace cleaver
     Json::Value root(Json::objectValue);
     root["name"] = "CREATE_FACE";
     root["id"] = catIds({ v1->tm_v_index, v2->tm_v_index, v3->tm_v_index}).c_str();
-    root["v1"] = v1->tm_v_index;
-    root["v2"] = v2->tm_v_index;
-    root["v3"] = v3->tm_v_index;
+    root["v1"] = static_cast<int>(v1->tm_v_index);
+    root["v2"] = static_cast<int>(v2->tm_v_index);
+    root["v3"] = static_cast<int>(v3->tm_v_index);
 
     if (face->triple && face->triple->order() == Order::TRIP) {
       root["triple"] = createVertexOperation(face->triple);   // TODO(jonbronson): set a UID and add operation to create separately
@@ -177,15 +177,15 @@ namespace cleaver
 
     Json::Value root(Json::objectValue);
     root["name"] = "CREATE_TET";
-    root["id"] = tet->tm_index;
+    root["id"] = static_cast<int>(tet->tm_index);
     root["verts"] = Json::Value(Json::arrayValue);
     for (int v = 0; v < VERTS_PER_TET; v++) {
-      root["verts"].append(verts[v]->tm_v_index);
+      root["verts"].append(static_cast<int>(verts[v]->tm_v_index));
     }
 
     if (tet->quadruple && tet->quadruple->order() == Order::QUAD) {
       // TODO(jonbronson): set a UID and add operation to create separately
-      root["quadruple"] = createVertexOperation(tet->quadruple);   
+      root["quadruple"] = createVertexOperation(tet->quadruple);
     }
 
     operations.push_back(root);
@@ -201,17 +201,17 @@ namespace cleaver
 		HalfEdge *edges[6];
 		HalfFace *faces[4];
 		mesh->getAdjacencyListsForTet(tet, verts, edges, faces);
-		
+
 		// create this tet - marked as debug tet
 		auto operations = createTetOperations(tet, mesh, true /* debug */);
 
-		// also create tets incident to each of the 4 vertices. 
+		// also create tets incident to each of the 4 vertices.
 		for (unsigned int v = 0; v < VERTS_PER_TET; v++) {
 			auto tets = mesh->tetsAroundVertex(verts[v]);
 			for (auto incidentTet : tets) {
 				if (incidentTet != tet) {
           auto tetOperations = createTetOperations(incidentTet, mesh);
-          operations.insert(operations.end(), tetOperations.begin(), tetOperations.end());					
+          operations.insert(operations.end(), tetOperations.begin(), tetOperations.end());
 				}
 			}
 		}
@@ -220,14 +220,14 @@ namespace cleaver
 	}
 
   Json::Value createVertexSnapOperation(
-    Vertex *vertex, const vec3 &warp_point, 
+    Vertex *vertex, const vec3 &warp_point,
     std::vector<HalfEdge*> violating_cuts,  std::vector<HalfEdge*> projected_cuts,
     std::vector<HalfFace*> violating_trips, std::vector<HalfFace*> projected_trips,
     std::vector<Tet*>      violating_quads, std::vector<Tet*>      projected_quads){
 
     Json::Value root(Json::objectValue);
     root["name"] = "SNAP_VERTEX";
-    root["vertex"] = vertex->tm_v_index;
+    root["vertex"] = static_cast<int>(vertex->tm_v_index);
     root["warp_point"] = Json::Value(Json::objectValue);
     root["warp_point"]["x"] = warp_point.x;
     root["warp_point"]["y"] = warp_point.y;
@@ -238,21 +238,21 @@ namespace cleaver
     for (HalfEdge *edge : violating_cuts) {
       root["violating_cuts"].append(idForEdge(edge).c_str());
     }
-    root["projected_cuts"] = Json::Value(Json::arrayValue);    
+    root["projected_cuts"] = Json::Value(Json::arrayValue);
     for (HalfEdge *edge : projected_cuts) {
       Json::Value cut = Json::Value(Json::objectValue);
       cut["id"] = idForEdge(edge).c_str();
-      cut["position"] = Json::Value(Json::objectValue);      
+      cut["position"] = Json::Value(Json::objectValue);
       cut["position"]["x"] = edge->cut->pos_next().x;
       cut["position"]["y"] = edge->cut->pos_next().y;
-      cut["position"]["z"] = edge->cut->pos_next().z;      
+      cut["position"]["z"] = edge->cut->pos_next().z;
       root["projected_cuts"].append(cut);
     }
 
     // add involved triples
     root["violating_triples"] = Json::Value(Json::arrayValue);
     for (HalfFace *face : violating_trips) {
-      root["violating_triples"].append(idForFace(face).c_str());      
+      root["violating_triples"].append(idForFace(face).c_str());
     }
     root["projected_triples"] = Json::Value(Json::arrayValue);
     for (HalfFace *face : projected_trips) {
@@ -261,7 +261,7 @@ namespace cleaver
       triple["position"] = Json::Value(Json::objectValue);
       triple["position"]["x"] = face->triple->pos_next().x;
       triple["position"]["y"] = face->triple->pos_next().y;
-      triple["position"]["z"] = face->triple->pos_next().z;      
+      triple["position"]["z"] = face->triple->pos_next().z;
       root["projected_triples"].append(triple);
     }
 
@@ -279,9 +279,9 @@ namespace cleaver
       quadruple["position"]["x"] = tet->quadruple->pos_next().x;
       quadruple["position"]["y"] = tet->quadruple->pos_next().y;
       quadruple["position"]["z"] = tet->quadruple->pos_next().z;
-      root["projected_quadruples"].append(quadruple);      
+      root["projected_quadruples"].append(quadruple);
     }
-    
+
     return root;
   }
